Add rutas para inmuebles
