+++ apps/client/src/sections/inmueble/view/index.js
@@ -1,1 +1,2 @@
 export { default as InmuebleView } from './inmueble-view';
+export { default as ListaInmuebleView } from './lista-inmuebles-view';
